backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# ── Deep Learning ─────────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
# Install CUDA variant instead for GPU:
# torch==2.3.0+cu121  torchvision==0.18.0+cu121  (from https://download.pytorch.org/whl/cu121)

# ── XAI Libraries ─────────────────────────────────────────────────────────────
grad-cam==1.5.0          # pytorch-grad-cam
lime==0.2.0.1
captum==0.7.0

# ── Image Processing ──────────────────────────────────────────────────────────
Pillow==10.3.0
opencv-python-headless==4.9.0.80
scikit-image==0.21.0
numpy==1.24.4

# ── Utilities ─────────────────────────────────────────────────────────────────
scipy==1.10.1
